chore: rename TS script for controling admission/discharge dates
