+++ ycms/static/src/index.ts
@@ -19,7 +19,7 @@ import { createIconsAt } from "./js/utils/create-icons";
 import "./js/patients-list";
 import "./js/utils/loader";
 import "./js/utils/set-language";
-import "./js/patient-intake";
+import "./js/admission-discharge-date-control";
 import "./js/specializations-list";
 
 window.addEventListener("DOMContentLoaded", () => {
